Yang models for system testing low level MD-SAL operations
Not every operation is easily available from Restconf.
This defines and agent application to control the tested operations from Java.

There are FIXMEs and TODOs left, to be fixed later,
probably together with Java implementation.
Merging this early will allow Changes for Bugs 7796-7807
to be implemented independently (as opposed to forming a single chain).

Change-Id: Icb9616c156f56f8706a24abe31e5ac097a65cf73
Signed-off-by: Vratko Polak <vrpolak@cisco.com>