Update patch set 1
Patch Set 1:

Looking at this some more, I don't see any code path which would yield a null Modification passed to addModification. Also, looking at the traces in the bug 7901, I don't think the NPE emanated from a BatchedModifications sent from the EOS b/c of the "The total number of batched messages received 4 does not match the number sent 6" error that was propagated to the ConcurrentDOMDataBroker a few millis later. This seems related to the serialization NPE as the latter would result in a lost BatchedModifications message. So I think the BatchedModifications message emanated from the RemoteTransactionContext.

I think the serialization NPE is a result of a modification of the list in BatchedModifications by the RemoteTransactionContext while writeExternal is iterating. This should cause a ConcurrentModException but I think it's possible it might not due to timing or thread visibility. In RemoteTransactionContext #sendBatchedModifications:

            sent = executeOperationAsync(batchedModifications,     
                             actorContext.getTransactionCommitOperationTimeout());
            if (ready) {
                batchedModifications = null;
            } else {
                batchedModifications = newBatchedModifications();
            }

If the client accessed the transaction concurrently, it's possible for the batchedModifications to reach akka remoting for serialization before batchedModifications is re-assigned after the call to executeOperationAsync and, in between, another modification operation is invoked to add another instance to batchedModifications. The RemoteTransactionContext is not thread-safe as it assumed clients will not access it concurrently.

Patch-set: 1