Bug 7814: Make transaction ID counter static
When a read tx actor os created in the Shard, the actor name is generated
as eg "shard-member-1:datastore-operational@0:19" where 0 is the instance's
static generation id and 19 is the tx id. So every part of the name is
constant for the controller instance except the tx id which is generated
via a counter in AbstractTransactionContextFactory. The problem is that
the AbstractTransactionContextFactory uses a class member for the tx id
counter and multiple AbstractTransactionContextFactory instances can be
create per datastore (one for regular transactions and one per tx chain).
Therefore a tx id may not be unique in the controller instance. This was
changed in Boron - prior to that the tx id counter was static so we didn't
see this problem. I changed it back to a static counter.

Change-Id: Iacb11bb401bd6bded38847690f8009c115ee0637
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>