Update patch set 1
Patch Set 1:

> Based on your description, the AbstractTransactionContextFactories
 > should differ in their historyId (0 for the normal transactions,
 > per-chain incrementing ID for the per-chain ones).
 > 
 > I wonder if we are failing to shut down an actor when we become
 > inactive or something ...

Hmm...yeah missed the historyId part. That differentiates IDs from the various AbstractTransactionContextFactories w/o needing a static tx id counter and the nextHistory counter starts at 1 so it doesn't conflict with 0 for normal transactions. The problem we had before with this was when the controller restarted but the persisted/incremented client ID on startup takes care of that.  It could be that the sal-distributed-datastore bundle is restarting during startup but that would/should generate a new client ID.

Patch-set: 1