Update patch set 1
Patch Set 1: Code-Review-1

Why is this needed? I think karaf4 feature service has already been fixed, and for karaf3 the problem is that we are racing with its internals -- i.e. feature:install (into which this class is not added) doing things while we are accessing it.

As Martin noted, an explanation is needed..

Patch-set: 1
Label: Code-Review=-1