Update patch set 3
Patch Set 3:

To be clear: Binding/NormalizedNode interactions have always a meeting point where we have both the binding class and the SchemaNode available. The details of that interaction are evaluated exactly once and the result is cached so we do not perform it more than once.

This place is a hot path and does not include any caches -- hence it is the wrong place to look at.

Patch-set: 3