BUG-5280: fix invalid local transaction replay
When we transition from a connecting to connected local
connection, we may encounter operations which are invalid
and these violations are detected during transaction replay.

If such replay fails, we need to suppress reporting the error
until the user initiates canCommit or directCommit, at which
point we need to report the delayed failure.

Change-Id: I2018498afff0e463dbdceaec5c50e8ebf088001b
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>