Update patch set 7
Patch Set 7:

> I thought what this patch did is the same thing as "couple simple
 > helper classes". The core implement of this patch was similar with
 > install-snapshot chunk mechanism. Other codes just try to make it
 > be more common and more easy to use.

Well this is a 5500 line patch with many interfaces, an apparent SPI/API,  several actors, a serialization component, a JMX component etc. That seems much more than a "couple simple helper classes" to me :)

I would prefer to start simple. To me the immediate problem is that we need to apply the similar raft snapshot chunking to remote read replies - w/o this we cannot move to artery. I'd like to see us focus on that with the new tell-based protocol. It already has actors in place - basically I think we need the equivalent of reusable LeaderInstallSnapshotState and SnapshotTracker classes and incorporate them into the existing tell-based FE/BE. So I would prefer to see a patch with just that to start.

Patch-set: 7