Update patch set 7
Patch Set 7:

> I am a big fan of keeping archetypes in top shape.
 > Is this Boron contribution salvageable?

Yes I don't see why not, but it would need a bit of TLC - rebase and possibly some adjustments (if memory from 1 year ago serves me right, it almost but not quite worked...).  Please feel free to jump on it! I'm a little under water to do this today - maybe I'll be able to pick it up some time in the coming weeks.

Patch-set: 7