Update patch set 1
Patch Set 1:

This looks OK. I've also been working on my own draft patch with reusable classes for just the message slicing and re-assembly. I'm designing it so it will fit in nicely with the new client FE/BE framework and could also be used for raft.  I've been incorporating parts of your classes as well.

Patch-set: 1