Update patch set 1
Patch Set 1:

> Well, this patch could be reusable for both  ReadData with request
 > & post and InstallSnapshot with post interface only.

We already have actors on both sides for both frameworks - I just need reusable slicing/re-assembly classes that can be easily integrated into the existing frameworks.  The classes also need to avoid exceeding the byte [] length limit (I introduced the FileBackedOutputStream for that earlier).

We might have a need for a separate channel framework with request/post at some point but not for the current uses cases.

Patch-set: 1