Update patch set 1
Patch Set 1:

> > I've already done that in my draft patch.
 > Ok. Thank you again?I would like  to continue to contribute to the
 > project as best as I can.

Yes - certainly. I don't want to make it seem like I'm hijacking or rejecting your work - I had intended to implement chunking in the new FE/BE framework for Nitrogen so we can switch to artery. I've looked at what you did as a starting point wrt to the slicing/re-assembly parts but am also designing it to take into account how the FE/BE framework would use it.

The separate channel with request/post is intriguing and we might need that for something in the future so let's re-visit that a bit later.

Patch-set: 1