Update patch set 1
Patch Set 1:

> > > I've already done that in my draft patch.
 > > Ok. Thank you again?I would like  to continue to contribute to
 > the
 > > project as best as I can.
 > 
 > Yes - certainly. I don't want to make it seem like I'm hijacking or
 > rejecting your work - I had intended to implement chunking in the
 > new FE/BE framework for Nitrogen so we can switch to artery. I've
 > looked at what you did as a starting point wrt to the
 > slicing/re-assembly parts but am also designing it to take into
 > account how the FE/BE framework would use it.
 > 
 > The separate channel with request/post is intriguing and we might
 > need that for something in the future so let's re-visit that a bit
 > later.

I knew, it's so kind of you. Actually class SubChannelProxy  is optional for the scene there is only an actor on one side and the other's not. And class SubChannel is the pure slicing/re-assembly part-:
Actor1--SubChannel <--post-------post---> SubChannel2--Actor2.
I've been thinking about the artery, it seems like it could only solve the akka cluster's heartbeats itself, but for our usecases, raft hearbeats and hugedata read  are the same thing for akka, since artery only decides channel based on dest-actor?

Patch-set: 1