Update patch set 1
Patch Set 1:

> I've been thinking about the artery, it seems like it could only
 > solve the akka cluster's heartbeats itself, but for our usecases,
 > raft hearbeats and hugedata read  are the same thing for akka,
 > since artery only decides channel based on dest-actor?

Artery is advertised to have a lot of improvements/advantages over the tcp framework, one is separate message channels as you mention. The dedicated channel for cluster heartbeats will definitely help us. We could use the large message channel for the FE/BE and the normal one for raft or vice versa.  I also recall reading a channel can have multiple parallel "lanes" for serialization but that feature isn't ready for production yet.

Patch-set: 1