Update patch set 1
Patch Set 1:

> I feel a bit confused about FE/BE,  does it need another separate
 > actor only to send and chunk large message ? and for normal
 > heartbeats message ,it could still use the original actor like
 > RaftActor. In another word, is it possible for RaftActor to send
 > both heartbeats and huge data to remote RaftActor using artery? or
 > must create a totally new actor  specifically for sending huge data
 > ?

FE/BE refers to the shard's client front-end framework that sends transaction-related messages to the shard back-end. This used to originate in the TransactionProxy class and was ask-based but Robert has re-implemented the whole front-end to use tell. There is a single FE actor for each data store that handles the replies. This is completely separate from RaftActor.

We can't isolate "huge" messages to any one actor. Any message containing a NormalizedNode could be "huge" and need to be chunked, including write/merge transaction requests sent from the FE, read replies sent from the shard BE, AppendEntries sent by RaftActor.  What I was saying is that we could use the 2 akka channels to separate RaftActor and client FE/BE traffic. However one would not be "small" and the other  "huge" - they would both have to be configured similarly with max frame size set to, say, 2M.

Patch-set: 1