Update patch set 1
Patch Set 1:

> I get it about FE/BE and agree that every message could be
 > "huge",so that we should use large message channel between FE and
 > BE actor, right?  My question is aout RaftActor- as AppendEntries
 > also work as heartbeats,  then how could we handle that when a
 > AppendEntry is huge and take long time, it would suspend the
 > heartbeat ,wouldn't it? Then how does RaftActor handle this ?

It's possible for an AooendEntries to be "huge" although the current max frame size is ~400M and I haven't heard of anyone hitting that. The cost of serializing the payload to a byte[] is charged to the RaftActor/Shard prior to appending to the log. This  could delay the heartbeat but shouldn't cause an election timeout in the follower b/c logic was added to query akka's cluster state to check if the leader node is actually down.

Patch-set: 1
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0