Bug 8424: Don't output data tree and tree candidates wih debug
Data trees and tree candidates can get quite large and fill up the
log files when debug is on or, worst case, cause OOM errors. For
debug logging, only print the whole tree/candidate with tract logging.
In cases where an Optional data tree is outputted via toString, only
output if the data tree is present or not.

Change-Id: I6cb5f9a5da9e3cc3218c83bb103b673db0fb1d80
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>