sal-binding-api: use lambdas
This series of patches uses lambdas instead of anonymous classes for
functional interfaces when possible. Lambdas are replaced with method
references when appropriate.

This restricts DataChangeListener’s type parameters to match its
parent’s, allowing it to be treated as a functional interface.
DataChangeListener is no longer used anywhere in ODL so the change
should be safe for Nitrogen. (I’ll follow up with a patch to remove
DataChangeListener.)

Change-Id: I57045a8cfb4dc788eb0d687eceb49ca54e6dc7c8
Signed-off-by: Stephen Kitt <skitt@redhat.com>