Update patch set 1
Patch Set 1:

I don''t think we should add anything in the factory conf as an example as that isn't supposed to be modified by users. This should be put in the user-facing akka.conf but commented out as we do for other settings. CSIT can set them appropriately in its akka.conf (I assume it already copies in its own with seed-nodes etc set appropriately).

Patch-set: 1