Update patch set 2
Patch Set 2: Code-Review-1

I'm wary of exposing a "backdoor" to just shut down the shard actor in the ShardManager  (without removing the replica) just for tests. I don't see a use case for it in production.  I think the test code should just issue the graceful Shutdown. It will leave a remnant in the ShardManager for the remainder of the test but that *should* be benign.

Patch-set: 2
Label: Code-Review=-1