Update patch set 2
Patch Set 2: -Code-Review

I think the problem this tries to address is between Shard actor lifecycle and ShardManager. Essentially if we just shut the actor down by bypassing ShardManager, the actor will still be present in ShardManager accounting structures, hence requests for resolution will be wrong.

This is not a typical production use case, so the point about introducing APIs is fair. On the other hand we need to solve that inconsistency somehow -- and I think the best avenue for that would be for ShardManager to have a deathwatch listener for Shard and react to it being shut down.

That way the test code can issue a clean shutdown, as proposed and ShardManager will notice (and it should holler loudly about that).

Is that agreeable?

Patch-set: 2
Reviewer: Gerrit User 649 <649@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0